% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepeak.R
\name{basepeak_comparison}
\alias{basepeak_comparison}
\title{Plot Basepeak of Samples Compared to Reference}
\usage{
basepeak_comparison(samples_df, reference_sample, n_plot_samples = 12)
}
\arguments{
\item{samples_df}{a tibble containing samplename and filename}

\item{reference_sample}{path to a reference sample}

\item{n_plot_samples}{number of samples to plot}
}
\description{
Plot Basepeak of Samples Compared to Reference
}
