% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamqc.R, R/utils.R
\docType{package}
\name{clamqc}
\alias{clamqc}
\title{\code{clamqc} package}
\description{
Calico Lipidomics and Metabolomics QC-associated R functions features methods for working with .mzrollDB mass spec datasets.
}
