% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepeak.R
\name{mzR_basepeak}
\alias{mzR_basepeak}
\title{Basepeak}
\usage{
mzR_basepeak(ms_file_path)
}
\arguments{
\item{ms_file_path}{Path to an mzXML, mzML, netCDF or mzData file}
}
\value{
a tibble with one row per MS1 scan and scan, rt, ion count columns
}
\description{
Find the largest ion count for each MS1 scan
}
