% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_plotting.R
\name{qc_standard_plots}
\alias{qc_standard_plots}
\title{QC Standard Plots}
\usage{
qc_standard_plots(standard_data)
}
\arguments{
\item{standard_data}{produced by \code{\link{qc_extract_standards}}}
}
\value{
a tibble containing standard QC grobs
}
\description{
QC Standard Plots
}
