% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards_mzR.R
\name{qc_standard_summaries}
\alias{qc_standard_summaries}
\title{QC Standard Summaries}
\usage{
qc_standard_summaries(extracted_standards)
}
\arguments{
\item{extracted_standards}{Output of \link{qc_extract_standards}}
}
\value{
a table of sample ~ standard ion counts
}
\description{
QC Standard Summaries
}
