% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards_mzrollDB.R
\name{standard_qc_peakgroups}
\alias{standard_qc_peakgroups}
\title{Standard QC for Peakgroups}
\usage{
standard_qc_peakgroups(
  mzroll_db_con,
  clamr_config,
  mass_spec_standards_con,
  query_methodId
)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances}

\item{mass_spec_standards_con}{A connection to an SQL database which stores
experimental mass spectrometry data.}

\item{query_methodId}{the analytical method being used}
}
\value{
the most plausible peakgroup label for each standard
}
\description{
Standard QC for Peakgroups
}
